update to version 1.5.1
mainly focus on the discovery of ultra-large variants

+++ src/RefSeqLoader.h
@@ -15,7 +15,7 @@ class RefSeqLoader {
 		string reg;
 		faidx_t *fai;
 		char *refseq;
-		int refseq_len;
+		int64_t refseq_len;
 
 	public:
 		RefSeqLoader(string &reg, faidx_t *fai);
